Organize components by type.
