+++ ComponentLib/CMakeLists.txt
@@ -60,14 +60,6 @@
 #    - Cameron Rutherford <cameron.rutherford@pnnl.gov>
 #]]
 
-add_subdirectory(Branch)
-add_subdirectory(Bus)
-add_subdirectory(Generator)
-add_subdirectory(Generator2)
-add_subdirectory(Generator4)
-add_subdirectory(Generator4Governor)
-add_subdirectory(Generator4Param)
-add_subdirectory(Load)
-add_subdirectory(MiniGrid)
+add_subdirectory(PowerFlow)
 add_subdirectory(PowerElectronics)
