+++ ComponentLib/DynamicPhasor/SynchronousMachine/README.md
