+++ ComponentLib/PowerFlow/Bus/BusFactory.hpp
@@ -57,112 +57,44 @@
  *
  */
 
+#pragma once
 
-#include <iostream>
-#include <cmath>
-#include <vector>
-#include "GeneratorPQ.hpp"
-#include <ComponentLib/Bus/BaseBus.hpp>
+#include <PowerSystemData.hpp>
+#include <ComponentLib/PowerFlow/Bus/BusPQ.hpp>
+#include <ComponentLib/PowerFlow/Bus/BusPV.hpp>
+#include <ComponentLib/PowerFlow/Bus/BusSlack.hpp>
 
 namespace ModelLib {
 
-/*!
- * @brief Constructor for a constant load model
- *
- * Calls default ModelEvaluatorImpl constructor.
- */
+    template <typename ScalarT = double, typename IdxT = int>
+    class BusFactory
+    {
+    public:
+        using real_type = typename ModelEvaluatorImpl<ScalarT, IdxT>::real_type;
+        using BusData = GridKit::PowerSystemData::BusData<real_type, IdxT>;
 
-template <class ScalarT, typename IdxT>
-GeneratorPQ<ScalarT, IdxT>::GeneratorPQ(bus_type* bus, GenData& data)
-  : P_(data.Pg),
-    Q_(data.Qg),
-    bus_(bus)
-{
-    //std::cout << "Create a load model with " << size_ << " variables ...\n";
-    size_ = 0;
-}
+        BusFactory() = delete;
 
-template <class ScalarT, typename IdxT>
-GeneratorPQ<ScalarT, IdxT>::~GeneratorPQ()
-{
-}
+        static BaseBus<ScalarT, IdxT>* create(BusData& data)
+        {
+            BaseBus<ScalarT, IdxT>* bus = nullptr;
+            switch(data.type)
+            {
+                case 1:
+                bus = new BusPQ<ScalarT, IdxT>(data);
+                break;
+                case 2:
+                bus = new BusPV<ScalarT, IdxT>(data);
+                break;
+                case 3:
+                bus = new BusSlack<ScalarT, IdxT>(data);
+                break;
+                default:
+                // Throw exception
+                std::cout << "Bus type " << data.type << " unrecognized.\n";
+            }
+            return bus;
+        }
+    };
 
-/*!
- * @brief allocate method computes sparsity pattern of the Jacobian.
- */
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::allocate()
-{
-    return 0;
-}
-
-/**
- * Initialization of the grid model
- */
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::initialize()
-{
-    return 0;
-}
-
-/*
- * \brief Identify differential variables
- */
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::tagDifferentiable()
-{
-    return 0;
-}
-
-/**
- * @brief Contributes to the bus residual.
- *
- * Must be connected to a PQ bus.
- */
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::evaluateResidual()
-{
-    // std::cout << "Evaluating load residual ...\n";
-    bus_->P() += P_;
-    bus_->Q() += Q_;
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::evaluateJacobian()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::evaluateIntegrand()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::initializeAdjoint()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::evaluateAdjointResidual()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int GeneratorPQ<ScalarT, IdxT>::evaluateAdjointIntegrand()
-{
-    return 0;
-}
-
-
-// Available template instantiations
-template class GeneratorPQ<double, long int>;
-template class GeneratorPQ<double, size_t>;
-
-
-} //namespace ModelLib
-
+} // namespace ModelLib
