+++ ComponentLib/PowerFlow/CMakeLists.txt
@@ -1,0 +1,14 @@
+# [[
+#  Author(s):
+#    - Cameron Rutherford <cameron.rutherford@pnnl.gov>
+#]]
+
+add_subdirectory(Branch)
+add_subdirectory(Bus)
+add_subdirectory(Generator)
+add_subdirectory(Generator2)
+add_subdirectory(Generator4)
+add_subdirectory(Generator4Governor)
+add_subdirectory(Generator4Param)
+add_subdirectory(Load)
+add_subdirectory(MiniGrid)
