+++ ComponentLib/PowerFlow/Generator/GeneratorFactory.hpp
@@ -57,126 +57,46 @@
  *
  */
 
+#pragma once
 
-#include <iostream>
-#include <cmath>
-#include <vector>
-#include "Load.hpp"
-#include <ComponentLib/Bus/BaseBus.hpp>
+#include <PowerSystemData.hpp>
+#include <ComponentLib/PowerFlow/Bus/BaseBus.hpp>
+#include <ComponentLib/PowerFlow/Generator/GeneratorSlack.hpp>
+#include <ComponentLib/PowerFlow/Generator/GeneratorPQ.hpp>
+#include <ComponentLib/PowerFlow/Generator/GeneratorPV.hpp>
+
 
 namespace ModelLib {
 
-/*!
- * @brief Constructor for a constant load model
- *
- * Calls default ModelEvaluatorImpl constructor.
- */
+    template <typename ScalarT = double, typename IdxT = int>
+    class GeneratorFactory
+    {
+    public:
+        using real_type = typename ModelEvaluatorImpl<ScalarT, IdxT>::real_type;
+        using GenData = GridKit::PowerSystemData::GenData<real_type, IdxT>;
 
-template <class ScalarT, typename IdxT>
-Load<ScalarT, IdxT>::Load(bus_type* bus, ScalarT P, ScalarT Q)
-  : P_(P),
-    Q_(Q),
-    busID_(0),
-    bus_(bus)
-{
-    //std::cout << "Create a load model with " << size_ << " variables ...\n";
-    size_ = 0;
-}
+        GeneratorFactory() = delete;
 
-template <class ScalarT, typename IdxT>
-Load<ScalarT, IdxT>::Load(bus_type* bus, LoadData& data)
-  : P_(data.Pd),
-    Q_(data.Qd),
-    busID_(data.bus_i),
-    bus_(bus)
-{
-    //std::cout << "Create a load model with " << size_ << " variables ...\n";
-    size_ = 0;
-}
+        static GeneratorBase<ScalarT, IdxT>* create(BaseBus<ScalarT, IdxT>* bus, GenData& data)
+        {
+            GeneratorBase<ScalarT, IdxT>* gen = nullptr;
+            switch(bus->BusType())
+            {
+                case 1:
+                gen = new GeneratorPQ<ScalarT, IdxT>(bus, data);
+                break;
+                case 2:
+                gen = new GeneratorPV<ScalarT, IdxT>(bus, data);
+                break;
+                case 3:
+                gen = new GeneratorSlack<ScalarT, IdxT>(bus, data);
+                break;
+                default:
+                // Throw exception
+                std::cout << "Generator type " << bus->BusType() << " unrecognized.\n";
+            }
+            return gen;
+        }
+    };
 
-template <class ScalarT, typename IdxT>
-Load<ScalarT, IdxT>::~Load()
-{
-}
-
-/*!
- * @brief allocate method computes sparsity pattern of the Jacobian.
- */
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::allocate()
-{
-    return 0;
-}
-
-/**
- * Initialization of the grid model
- */
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::initialize()
-{
-    return 0;
-}
-
-/*
- * \brief Identify differential variables
- */
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::tagDifferentiable()
-{
-    return 0;
-}
-
-/**
- * @brief Contributes to the bus residual.
- *
- * Must be connected to a PQ bus.
- */
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::evaluateResidual()
-{
-    // std::cout << "Evaluating load residual ...\n";
-    bus_->P() -= P_;
-    bus_->Q() -= Q_;
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::evaluateJacobian()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::evaluateIntegrand()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::initializeAdjoint()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::evaluateAdjointResidual()
-{
-    return 0;
-}
-
-template <class ScalarT, typename IdxT>
-int Load<ScalarT, IdxT>::evaluateAdjointIntegrand()
-{
-    return 0;
-}
-
-
-
-
-// Available template instantiations
-template class Load<double, long int>;
-template class Load<double, size_t>;
-
-
-} //namespace ModelLib
-
+} // namespace ModelLib
