+++ ComponentLib/PowerFlow/Generator/GeneratorSlack.cpp
@@ -61,8 +61,8 @@
 #include <iostream>
 #include <cmath>
 #include <vector>
-#include "MiniGrid.hpp"
-#include <ComponentLib/Bus/BaseBus.hpp>
+#include "GeneratorSlack.hpp"
+#include <ComponentLib/PowerFlow/Bus/BaseBus.hpp>
 
 namespace ModelLib {
 
@@ -73,26 +73,15 @@ namespace ModelLib {
  */
 
 template <class ScalarT, typename IdxT>
-MiniGrid<ScalarT, IdxT>::MiniGrid()
-  : ModelEvaluatorImpl<ScalarT, IdxT>(3, 0, 0),
-    Pl2_(  2.5),
-    Ql2_( -0.8),
-    Pg3_(  2.0),
-    V1_ (  1.0),
-    th1_(  0.0),
-    V3_ (  1.1),
-    B12_( 10.0),
-    B13_( 15.0),
-    B22_(-22.0),
-    B23_( 12.0)
+GeneratorSlack<ScalarT, IdxT>::GeneratorSlack(bus_type* bus, GenData& data)
+  : bus_(bus)
 {
     //std::cout << "Create a load model with " << size_ << " variables ...\n";
-    rtol_ = 1e-5;
-    atol_ = 1e-5;
+    size_ = 0;
 }
 
 template <class ScalarT, typename IdxT>
-MiniGrid<ScalarT, IdxT>::~MiniGrid()
+GeneratorSlack<ScalarT, IdxT>::~GeneratorSlack()
 {
 }
 
@@ -100,7 +89,7 @@ MiniGrid<ScalarT, IdxT>::~MiniGrid()
  * @brief allocate method computes sparsity pattern of the Jacobian.
  */
 template <class ScalarT, typename IdxT>
-int MiniGrid<ScalarT, IdxT>::allocate()
+int GeneratorSlack<ScalarT, IdxT>::allocate()
 {
     return 0;
 }
@@ -109,14 +98,19 @@ int MiniGrid<ScalarT, IdxT>::allocate()
  * Initialization of the grid model
  */
 template <class ScalarT, typename IdxT>
-int MiniGrid<ScalarT, IdxT>::initialize()
+int GeneratorSlack<ScalarT, IdxT>::initialize()
 {
-    th2() = 0.0; // th2
-    V2()  = 1.0; // V2
-    th3() = 0.0; // th3
     return 0;
 }
 
+/*
+ * \brief Identify differential variables
+ */
+template <class ScalarT, typename IdxT>
+int GeneratorSlack<ScalarT, IdxT>::tagDifferentiable()
+{
+    return 0;
+}
 
 /**
  * @brief Contributes to the bus residual.
@@ -124,24 +118,50 @@ int MiniGrid<ScalarT, IdxT>::initialize()
  * Must be connected to a PQ bus.
  */
 template <class ScalarT, typename IdxT>
-int MiniGrid<ScalarT, IdxT>::evaluateResidual()
+int GeneratorSlack<ScalarT, IdxT>::evaluateResidual()
 {
-    f_[0] = -Pl2_ - V2()*(V1_*B12_*sin(th2()-th1_) + V3_*B23_*sin(th2() - th3()));
-    f_[1] = -Ql2_ + V2()*(V1_*B12_*cos(th2()-th1_) + B22_*V2() + V3_*B23_*cos(th2() - th3()));
-    f_[2] =  Pg3_ - V3_ *(V1_*B13_*sin(th3()-th1_) + V2()*B23_*sin(th3() - th2()));
-
+    // std::cout << "Evaluating load residual ...\n";
+    // bus_->P() += P_;
+    // bus_->Q() += Q_;
     return 0;
 }
 
 template <class ScalarT, typename IdxT>
-int MiniGrid<ScalarT, IdxT>::evaluateJacobian()
+int GeneratorSlack<ScalarT, IdxT>::evaluateJacobian()
 {
     return 0;
 }
 
+template <class ScalarT, typename IdxT>
+int GeneratorSlack<ScalarT, IdxT>::evaluateIntegrand()
+{
+    return 0;
+}
+
+template <class ScalarT, typename IdxT>
+int GeneratorSlack<ScalarT, IdxT>::initializeAdjoint()
+{
+    return 0;
+}
+
+template <class ScalarT, typename IdxT>
+int GeneratorSlack<ScalarT, IdxT>::evaluateAdjointResidual()
+{
+    return 0;
+}
+
+template <class ScalarT, typename IdxT>
+int GeneratorSlack<ScalarT, IdxT>::evaluateAdjointIntegrand()
+{
+    return 0;
+}
+
+
+
+
 // Available template instantiations
-template class MiniGrid<double, long int>;
-template class MiniGrid<double, size_t>;
+template class GeneratorSlack<double, long int>;
+template class GeneratorSlack<double, size_t>;
 
 
 } //namespace ModelLib
