+++ ComponentLib/PowerFlow/Generator4Param/static_data.hpp
